correction type chrono (double=>long)

+++ BattleCalculator.Api/Models/Game/CreateGameResponse.cs
@@ -12,7 +12,7 @@ namespace BattleCalculator.Api.Models.Game
 
 		public int Level { get; set; }
 
-		public double Chrono { get; set; }
+		public long Chrono { get; set; }
 
 	}
 }
