+++ BattleCalculator.Model/Entities/Game.cs
@@ -8,7 +8,7 @@ namespace BattleCalculator.Model.Entities
     {
         public int Id { get; set; }
         public int Level { get; set; }
-        public double Chrono { get; set; }
+        public long Chrono { get; set; }
         public int TotalScore { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime EndedAt { get; set; }
